# BackEnd/server-master/server-master/api.yml
openapi: 3.0.0
info:
  title: Products API
  description: API for managing products and cart total
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /products:
    post:
      summary: Add a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                price:
                  type: number
                type:
                  type: string
      responses:
        200:
          description: Product added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    get:
      summary: Get all products or filter/sort them
      parameters:
        - name: filter
          in: query
          required: false
          schema:
            type: string
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [price]
      responses:
        200:
          description: List of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        title:
                          type: string
                        price:
                          type: number
                        type:
                          type: string
  /products/{id}:
    put:
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                price:
                  type: number
                type:
                  type: string
                selected:
                  type: boolean
      responses:
        200:
          description: Product updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    delete:
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Product deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /product/{id}:
    get:
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Product details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      title:
                        type: string
                      price:
                        type: number
                      type:
                        type: string
  /cart/total:
    get:
      summary: Get total amount of the cart
      responses:
        200:
          description: Total cart amount
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: number